Tue, Feb 28, 2023, 10:08 PM +06:00

+++ beecrowd/1973.c
@@ -14,20 +14,20 @@ then,
 
 int main(void)
 {
-    unsigned long long int n;
-    scanf("%llu", &n);
+    unsigned int n;
+    scanf("%u", &n);
 
-    unsigned long long int sheep[n], ref[n];
+    unsigned int sheep[n], ref[n];
     //taking sheep value and an identical copy
-    for (unsigned long long int i = 0; i < n; i++)
+    for (unsigned int i = 0; i < n; i++)
     {
-        scanf("%llu", &sheep[i]);
+        scanf("%u", &sheep[i]);
         ref[i] = sheep[i];
     }
 
-    unsigned long long int attacked = 0, non_stolen = 0;
+    unsigned int attacked = 0, non_stolen = 0;
 
-    for (unsigned long long int j = 0;;)
+    for (unsigned int j = 0;;)
     {
         //odd condision
         if (sheep[j] % 2 != 0)
@@ -51,7 +51,7 @@ int main(void)
     }
 
     // checking the number of attacked stars
-    for (unsigned long long int i = 0; i < n; i++)
+    for (unsigned int i = 0; i < n; i++)
     {
         if (sheep[i] != ref[i])
         {
@@ -60,12 +60,12 @@ int main(void)
     }
 
     // checking the number of remaining sheeps
-    for (unsigned long long int i = 0; i < n; i++)
+    for (unsigned int i = 0; i < n; i++)
     {
         non_stolen += sheep[i];
     }
 
-    printf("%llu %llu\n", attacked, non_stolen);
+    printf("%u %u\n", attacked, non_stolen);
 
     return 0;
 }
